% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBlockH2.R
\name{addBlockH2}
\alias{addBlockH2}
\title{Adds an H2 (heading two) Block to a Page}
\usage{
addBlockH2(secret, id, content, toggle = FALSE)
}
\description{
Id refers to a page id, content should be only text.
@param secret API token
@param id Page id where block will be appended
@param content content to append as H2
@param toggle defaults to FALSE. If TRUE, will create an H1 Toggle.
}
