% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBlockParagraph.R
\name{addBlockParagraph}
\alias{addBlockParagraph}
\title{Adds a Paragraph Block to a Page}
\usage{
addBlockParagraph(secret, id, content)
}
\description{
Id refers to a page id, content should be only text. HTML will export to text.
@param secret API token
@param id Page id where block will be appended
@param content content to append
}
