% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_page.r
\name{archive_page}
\alias{archive_page}
\title{Archive a Page}
\usage{
archive_page(secret, id)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{page id}
}
\value{
list of response
}
\description{
Archive's a notion page.
}
\author{
Eduardo Flores
}
