% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivePage.R
\name{archivePage}
\alias{archivePage}
\title{Archive a Page}
\usage{
archivePage(secret, id)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{page id}
}
\value{
list of response
}
\description{
Archive's a notion page.
}
\author{
Eduardo Flores
}
