% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{cached_access_code}
\alias{cached_access_code}
\title{Retrieve cached Notion access token}
\usage{
cached_access_code(path = NULL)
}
\arguments{
\item{path}{A path to the local directory that contains the Notion access
token. If left \code{NULL}, it will be set as the default cache directory or
search for the environment variable \code{NOTIONR_CACHE}.}
}
\value{
A character string that is the user's Notion access token.
If no access token exists, an error will be thrown
}
\description{
\code{cached_access_code} retrieves a cached Notion access token from the default
cache directory, directory defined by environment variable, or user defined path
}
