% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{children}
\alias{children}
\title{Retrieve the children of a block object}
\usage{
children(x, recursive = TRUE)
}
\arguments{
\item{x}{A page or block.}

\item{recursive}{A boolean value. If \code{TRUE}, this will recursively access
all children of the specified block. If \code{FALSE}, it will return only the
first level of block children.}
}
\value{
If recursive, a tree containing all block children. Otherwise a
single-level list of block children.
}
\description{
This function allows the user to recursively/non-recursively access a block
object's children. NOTE: Pages are also blocks and therefore may have
children as well.
}
