% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{compound_filter}
\alias{compound_filter}
\title{Compound database filter}
\usage{
compound_filter(x, y, type = "and")
}
\arguments{
\item{x}{A property filter; either compound or single-level.}

\item{y}{A property filter; either compound or single-level.}

\item{type}{A string indicating what type of compound filter. Must be either
\code{"and"} or \code{"or"}.}
}
\value{
A structured compound filter.
}
\description{
The \code{compound_filter} function combines two filters into a compound one. These
filters can either be compound filters themselves, or single-level filters as
described in \code{\link{property_filter}}. This filter is only meant for use
within the \code{\link{query_database}} function.
}
\details{
Compound filters can be nested up to 2 levels deep. Currently there is no
checking for this, so it's on the user to adhere to it.
}
\examples{
compound_filter(
  x = property_filter(
    property = "Landmark",
    type = "text",
    body = contains ~ Bridge
  ),
  y = property_filter(
    property = "Bodies of Water",
    type = "checkbox",
    body = equals ~ TRUE
  ),
  type = "and"
)
}
