% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{compound_sort}
\alias{compound_sort}
\title{Compound database sort}
\usage{
compound_sort(...)
}
\arguments{
\item{...}{A list of property sorts.}
}
\value{
An object of class \code{notionr_compund_sort}.
}
\description{
The \code{compound_sort} function combines a list of sorts into a compound one.
Each element in this list should be an output of \code{\link{property_sort}}.
This function is only meant for use
within the \code{\link{query_database}} function.
}
\examples{
compound_sort(
  property_sort(
    property = "Ingredients",
    timestamp = "last_edited_time",
    direction = "descending"
  ),
  property_sort(
    property = "Travel Destinations",
    timestamp = "created_time",
    direction = "ascending"
  )
)
}
