% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_notion_page.r
\name{create_notion_page}
\alias{create_notion_page}
\title{Create a Page}
\usage{
create_notion_page(secret, parent_id, title_property, title = "untitled")
}
\arguments{
\item{secret}{API token}

\item{parent_id}{database id where page will be created or page under which it will nest}

\item{title_property}{name of the title property}

\item{title}{title of the new page}
}
\description{
Silent Return
}
\author{
Eduardo Flores
}
