% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNotionPage.R
\name{createNotionPage}
\alias{createNotionPage}
\title{Create a Page}
\usage{
createNotionPage(secret, parent_id, title_property, title = "untitled")
}
\arguments{
\item{secret}{API token}

\item{parent_id}{database id where page will be created or page under which it will nest}

\item{title_property}{name of the title property}

\item{title}{title of the new page}
}
\value{
list
}
\description{
Silent Return
}
\author{
Eduardo Flores
}
