% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_block.r
\name{delete_block}
\alias{delete_block}
\title{Delete a Block}
\usage{
delete_block(secret, id)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{block id}
}
\value{
list of response
}
\description{
Deletes a block
}
\author{
Eduardo Flores
}
