% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteBlock.R
\name{deleteBlock}
\alias{deleteBlock}
\title{Delete a Block}
\usage{
deleteBlock(secret, id)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{block id}
}
\value{
list of response

list
}
\description{
Deletes a block
}
\author{
Eduardo Flores
}
