% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{details}
\alias{details}
\title{Print detailed object summary}
\usage{
details(x, ...)
}
\arguments{
\item{x}{A page or database.}

\item{...}{Additional arguments to pass to \code{format}.}
}
\value{
Invisibly returns itself. Called primarily for the side effects.
}
\description{
\code{details} will display a detailed summary for pages or databases
in a tree-like fashion. This is particularly helpful for getting a quick
overview of an object's structure.
}
