% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\alias{notion_filter}
\alias{notion_or}
\alias{add_checkbox_filter}
\alias{add_select_filter}
\title{Filter Operators}
\usage{
notion_filter(., sort = NULL)

notion_or(...)

add_checkbox_filter(property, equals = TRUE)

add_select_filter(property, equals)
}
\arguments{
\item{.}{List of filter conditions}

\item{sort}{List of sort conditions. NULL by default.}

\item{property}{name or id of property (column) in database}

\item{equals}{TRUE (default) or FALSE condition to meet in checkbox. Equals (contains) for select filter.}
}
\value{
list

list
}
\description{
Helps you build a database filter call with human-readable intuition. You must finish (if using pipes) with notion_filter().

Adds a checkbox filter condition.
}
\details{
Still WIP, only a few operators are currently available.
}
\examples{
# add a condition where checkbox should be checked
add_checkbox_filter("id_column")
# add a condition where checkbox should NOT be checked
add_checkbox_filter("id_column", FALSE)
}
\author{
Eduardo Flores

Eduardo Flores
}
