% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notion_page.r
\name{get_notion_page}
\alias{get_notion_page}
\title{Get a Page}
\usage{
get_notion_page(secret, id)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{page id}
}
\value{
list of response
}
\description{
Gets a Notion Page
}
\author{
Eduardo Flores
}
