% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notion_page_content.r
\name{get_notion_page_content}
\alias{get_notion_page_content}
\title{Get a Page with it's Content}
\usage{
get_notion_page_content(secret, id)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{page id}
}
\value{
list of response
}
\description{
Gets a Notion Page and the content inside it. Works with nested blocks too, as a page is also a block.
}
\author{
Eduardo Flores
}
