% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlock.R
\name{getBlock}
\alias{getBlock}
\title{Returns a block}
\usage{
getBlock(secret, id)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{Notion block ID.}
}
\value{
JSON
}
\description{
Query API with a block ID and retrieve JSON format (for now)
}
\author{
Eduardo Flores
}
