% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMultiSelectValues.R
\name{getMultiSelectValues}
\alias{getMultiSelectValues}
\title{Returns a vector of the available multi-select options in a field}
\usage{
getMultiSelectValues(
  column,
  strip_string = "\\\\|",
  no_na = TRUE,
  only_unique = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{column}{vector of column in data.frame (usually, after using getNotionDatabase, in form db$column_name )}

\item{strip_string}{String by which we should strip >1 selects in a single database row. Defaults to "|", as this is the default behaviour in getNotionDatabase().}

\item{no_na}{Strip all NA's? Defaults to TRUE.}

\item{only_unique}{Export only unique values in vector? Defaults to TRUE.}

\item{show_progress}{Print the count of values? Defaults to TRUE.}
}
\value{
vector
}
\description{
For a given column in a database field, creates a vector of all entries. See parameters for options.
}
\details{
does NOT call API.
}
\author{
Eduardo Flores
}
