% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNotionPage.R
\name{getNotionPage}
\alias{getNotionPage}
\title{Get a Page}
\usage{
getNotionPage(secret, id)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{page id}
}
\value{
list of response
}
\description{
Gets a Notion Page
}
\author{
Eduardo Flores
}
