% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNotionPageContent.R
\name{getNotionPageContent}
\alias{getNotionPageContent}
\title{Get a Page with it's Content}
\usage{
getNotionPageContent(secret, id)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{page id}
}
\value{
list of response
}
\description{
Gets a Notion Page and the content inside it. Works with nested blocks too, as a page is also a block.
}
\author{
Eduardo Flores
}
