% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{has_children}
\alias{has_children}
\title{Check if a block has children}
\usage{
has_children(x)
}
\arguments{
\item{x}{An object with class \code{notionr_block} or \code{notionr_page}}
}
\value{
Returns \code{TRUE} if its argument has children and \code{FALSE} otherwise.
}
\description{
Identifies if a block or page object has children.
}
