% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{id}
\alias{id}
\title{Retrieve page and database IDs}
\usage{
id(x)
}
\arguments{
\item{x}{A page or database.}
}
\value{
A data.frame containing the object's name and ID in a tidy format.
}
\description{
The \code{id} function returns the title and ID of pages and databases. In some
functions the object name and ID can be used interchangeably, although
using the ID is usually recommended.
}
