% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEmptyNotionDatabaseExport.R
\name{isEmptyNotionDatabaseExport}
\alias{isEmptyNotionDatabaseExport}
\title{Is the database output empty ?}
\usage{
isEmptyNotionDatabaseExport(dataframe)
}
\arguments{
\item{dataframe}{The database output}
}
\value{
vector
}
\description{
Is the database output empty ?
}
