% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{list_databases}
\alias{list_databases}
\title{List all Databases}
\usage{
list_databases(key, query = NULL, sort = NULL)
}
\arguments{
\item{key}{Notion access key as a character.}

\item{query}{A string which limits which databases are returned by comparing
the query to the page title. If \code{NULL}, no limiting occurs.}

\item{sort}{A search sort object. If \code{NULL}, no sorting will occur.}
}
\value{
A list of database objects.
}
\description{
Return a list of all database objects. Allows the user to apply a query
filter or sort direction, as seen in the \code{\link{search}} function, to
alter how/which objects are returned.
}
\seealso{
\code{\link[=search]{search()}} for examples of sorts.
}
