% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{list_page_ids}
\alias{list_page_ids}
\title{Retrieve all page names and IDs}
\usage{
list_page_ids(key, query = NULL, sort = NULL)
}
\arguments{
\item{key}{Notion access key as a character.}

\item{query}{A string which limits which pages are returned by comparing the
query to the page title. If \code{NULL}, no limiting occurs.}

\item{sort}{A search sort object. If \code{NULL}, no sorting will occur.}
}
\value{
A data.frame containing page names and IDs.
}
\description{
Access all page names and IDs in a tidy format. This is particularly
useful for getting a quick overview of all pages and selecting a
specific ID to access an individual page.
}
\seealso{
\code{\link[=search]{search()}} and \code{\link[=list_pages]{list_pages()}}
}
