% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{list_pages}
\alias{list_pages}
\title{List all Pages}
\usage{
list_pages(key, query = NULL, sort = NULL)
}
\arguments{
\item{key}{Notion access key as a character.}

\item{query}{A string which limits which pages are returned by comparing the
query to the page title. If \code{NULL}, no limiting occurs.}

\item{sort}{A search sort object. If \code{NULL}, no sorting will occur.}
}
\value{
A list of database objects.
}
\description{
Return a list of all page objects. Allows the user to apply a query
filter or sort direction, as seen in the \code{\link{search_workspace}} function, to
alter how/which objects are returned.
}
\seealso{
\code{\link[=search_workspace]{search_workspace()}} for examples of sorts.
}
