% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{list_users}
\alias{list_users}
\title{List all workspace users}
\usage{
list_users(key)
}
\arguments{
\item{key}{Notion access key.}
}
\value{
A data.frame with user information.
}
\description{
This function retrieves user information for all users with access to a
workspace and returns it in a tidy format.
}
