% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_chrome_page_ids.r
\name{normalize_chrome_page_ids}
\alias{normalize_chrome_page_ids}
\title{Returns a page id from a copy-paste page id in browser}
\usage{
normalize_chrome_page_ids(x)
}
\arguments{
\item{x}{page id in format: 1f4c70197b0e4589902d371adc1dbd9a}
}
\value{
character
}
\description{
Returns a page id from a copy-paste page id in browser
}
\author{
Eduardo Flores
}
