% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeChromePageIds.R
\name{normalizeChromePageIds}
\alias{normalizeChromePageIds}
\title{Returns a page id from a copy-paste page id in browser}
\usage{
normalizeChromePageIds(x)
}
\arguments{
\item{x}{page id in format: 1f4c70197b0e4589902d371adc1dbd9a}
}
\value{
character
}
\description{
Returns a page id from a copy-paste page id in browser
}
\author{
Eduardo Flores
}
