% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{object_content}
\alias{object_content}
\title{Extract object's primary content}
\usage{
object_content(x)
}
\arguments{
\item{x}{A page or block object.}
}
\value{
A list; this contains the content contained within all relevant
children or properties within a block or a page.
}
\description{
Pages and blocks contain different content. This function provides a
straightforward way to access the content from these relevant object types.
}
