% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich-text.R
\name{plain_text}
\alias{plain_text}
\title{Extract plain text from rich text object}
\usage{
plain_text(x)
}
\arguments{
\item{x}{A rich text object}
}
\value{
The plain text as a string.
}
\description{
A rich text object has both a content property and a plain-text property. To
access the content property, see \code{\link{object_content}}, and to access
the plain text, use this function.
}
