% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich-text.R
\name{plain_text_array}
\alias{plain_text_array}
\title{Extract plain text from a rich text array}
\usage{
plain_text_array(x)
}
\arguments{
\item{x}{A rich text array.}
}
\value{
A list of plain text elements.
}
\description{
This function extracts all plain text elements from a rich text array object.
}
\seealso{
\code{\link[=plain_text]{plain_text()}}
}
