% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{properties}
\alias{properties}
\title{Retrieve page and database properties}
\usage{
properties(x)
}
\arguments{
\item{x}{A page or database.}
}
\value{
A data.frame containing structured, high-level information about the
object's properties.
}
\description{
Database and page properties contain important structural information and
content. This function retrieves high-level information about these
properties and returns them in a tidy format.
}
