% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{property_filter}
\alias{property_filter}
\title{Filter condition for database query}
\usage{
property_filter(property, type, body)
}
\arguments{
\item{property}{A character string; the name of the property to apply the
filter to.}

\item{type}{A character string; the property type as identified by Notion.}

\item{body}{A two-sided formula. The left hand side specifies the filter
condition property - valid condition properties depend on the filter type -
and the right had side specifies the filtering value.}
}
\value{
A list; a formatted property filter.
}
\description{
The \code{property_filter()} function applies a filter condition within a database
query to limit which pages are returned.
}
\details{
The \code{property_filter()} applies to a particular database property, by name or
id. It applies a user-supplied condition to this property to limit which
pages are returned. For a full set of valid properties and property types,
see the \href{https://developers.notion.com/reference/post-database-query#post-database-query-filter}{documentation here}.
This filter is only meant for use within the \code{\link{query_database}}
function.
}
\examples{
property_filter(
  property = "Landmark",
  type = "rich_text",
  body = contains ~ "Bridge"
)
}
