% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{property_sort}
\alias{property_sort}
\title{Sort condition for database query}
\usage{
property_sort(
  property,
  timestamp = "last_edited_time",
  direction = "descending"
)
}
\arguments{
\item{property}{A character string; the name of the property to apply the
sort to.}

\item{timestamp}{A character string; must be one of \code{last_edited_time} or
\code{created_time}.}

\item{direction}{A character string; must be one of \code{descending} or \code{ascending}.}
}
\value{
A list; a formatted property sort.
}
\description{
The \code{property_sort()} function applies a sort condition within a database
query to sort how pages are returned.
}
\details{
The \code{property_sort()} function applies to a particular database property,
by name or id. It applies a user-supplied condition to this property to sort
pages are returned. This sort is only meant for use within the
\code{\link{query_database}} function.
}
\examples{
property_sort(
  property = "Ingredients",
  timestamp = "last_edited_time",
  direction = "descending"
)
}
