% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{query_body}
\alias{query_body}
\title{Construct body for database query}
\usage{
query_body(filter = NULL, sorts = NULL)
}
\arguments{
\item{filter}{A property filter object.}

\item{sorts}{A property sort or compound sort object.}
}
\value{
The formatted query body.
}
\description{
This function constructs the query body that is used to filter/sort the
results of the database query.
}
\seealso{
\code{\link[=property_filter]{property_filter()}} and \code{\link[=property_sort]{property_sort()}} to see how to properly
create property filters and sorts.
}
