% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{query_database}
\alias{query_database}
\title{Query a database}
\usage{
query_database(database.id, key, query.body = NULL)
}
\arguments{
\item{database.id}{The database ID as a character.}

\item{key}{Notion access key.}

\item{query.body}{The query body which contains potential filters and sorts.}
}
\value{
A list of pages.
}
\description{
Query a database and return pages contained in that database. Also allows the
user to apply sorting and filtering logic to customize which/how pages are
returned. For details on how to structure these filters and sorts, see
\code{\link{query_body}}, \code{\link{property_filter}}, and
\code{\link{property_sort}}.
}
