% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{retrieve_database}
\alias{retrieve_database}
\title{Retrieve an individual database}
\usage{
retrieve_database(key, database.id)
}
\arguments{
\item{key}{Notion access key as a string.}

\item{database.id}{Unique database identifier as a string.}
}
\value{
A database object.
}
\description{
Access a specific database using it's unique identifier.
}
