% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{retrieve_page}
\alias{retrieve_page}
\title{Retrieve an individual page}
\usage{
retrieve_page(key, page.id)
}
\arguments{
\item{key}{Notion access key as a string.}

\item{page.id}{Unique database identifier as a string.}
}
\value{
A page object.
}
\description{
Access a specific page using it's unique identifier.
}
