% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{retrieve_user}
\alias{retrieve_user}
\title{Retrieve a user object}
\usage{
retrieve_user(key, user.id)
}
\arguments{
\item{key}{Notion access key.}

\item{user.id}{Unique user ID as a string.}
}
\value{
A user object.
}
\description{
This function will retrieve a particular user, referenced by ID. This ID can
most easily be accessed using the \code{\link{list_users}} function.
}
