% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Search for databases and pages}
\usage{
search(key, query = NULL, sort = NULL, filter = NULL)
}
\arguments{
\item{key}{Notion access key.}

\item{query}{A string which limits which pages are returned by comparing the
query to the page title. If \code{NULL}, no limiting occurs.}

\item{sort}{A search sort object. If \code{NULL}, no sorting will occur.}

\item{filter}{A search filter object. If \code{NULL} no filtering will occur.}
}
\value{
A list of pages and/or databases.
}
\description{
Search for databases and pages, using filters and sorts to limit which/how
pages and databases are returned.
}
\seealso{
\code{\link[=search_filter]{search_filter()}} and \code{\link[=search_sort]{search_sort()}} to see details on how to
correctly construct filter and sort objects.
}
