% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_filter}
\alias{search_filter}
\title{Create a search filter}
\usage{
search_filter(value = "page", property = "object")
}
\arguments{
\item{value}{Which type of property to filter to. Currently, \code{"page"} and
\code{"database"} are the only valid options.}

\item{property}{Which property type to apply the filter to. Currently
\code{"object"} is the only valid option.}
}
\value{
A search filter object.
}
\description{
Construct a search filter to filter results returned from the \code{\link{search}}
function. This filter is only meant for use within the \code{\link{search}}
function.
}
