% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_sort}
\alias{search_sort}
\title{Create a search sort}
\usage{
search_sort(timestamp = "last_edited_time", direction = "descending")
}
\arguments{
\item{timestamp}{Which property to sort by. Currently, \code{"last_edited_time"}
is the only valid option.}

\item{direction}{Which direction to sort in. Must be either \code{"descending"} or
\code{"ascending"}.}
}
\value{
A search sort object.
}
\description{
Construct a search sort to sort results returned from the \code{\link{search}}
function. This sort is only meant for use within the \code{\link{search}}
function.
}
