% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{title}
\alias{title}
\title{Retrieve page or database title}
\usage{
title(x, all.titles = FALSE)
}
\arguments{
\item{x}{A page or database.}

\item{all.titles}{A boolean indicating whether to return only the first title
array element, or all elements.}
}
\value{
If \code{all.titles} is \code{TRUE}, a list of all title elements. Otherwise,
a single title value.
}
\description{
Pages and Database objects store their title property as an array of
\href{https://developers.notion.com/reference/rich-text}{rich text objects}.
Typically the title only consists of one relevant element, and thus this
function will return the first element as a default.
}
