% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_checkbox.r
\name{update_checkbox}
\alias{update_checkbox}
\title{Updates a Checkbox Property}
\usage{
update_checkbox(secret, id, property_name, value = TRUE)
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
@param secret API token
@param id Page id to be updated
@param property_name name of property to update (should be a checkbox type property)
@param value value to update. Use R bollean object. Defaults to TRUE.
}
