% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_date.r
\name{update_date}
\alias{update_date}
\title{Updates a Date property}
\usage{
update_date(secret, id, property_name, value)
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
@param secret API token
@param id Page id to be updated
@param property_name name of property to update (should be a date type property)
@param value value to update
}
