% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_emoji.r
\name{update_emoji}
\alias{update_emoji}
\title{Updates an Emoji Property of a page}
\usage{
update_emoji(secret, id, emoji)
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
@param secret API token
@param id Page id to be updated
@param emoji emoji to update to
}
