% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_page_cover.r
\name{update_page_cover}
\alias{update_page_cover}
\title{Updates a Page Cover}
\usage{
update_page_cover(secret, id, cover_url)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{page id to be updated}

\item{cover_url}{url of cover to be update}
}
\value{
list of response
}
\description{
Updates a page cover to the url specified
}
\author{
Eduardo Flores
}
