% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_relationship.r
\name{update_relationship}
\alias{update_relationship}
\title{Updates (adds) a relationship to a page id}
\usage{
update_relationship(secret, id, property_name, value)
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
@param secret API token
@param id Page id to be updated
@param property_name name of property to update (should be a relationship type property)
@param value value to update (should be a unique page id)
}
