% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_text.r
\name{update_text}
\alias{update_text}
\title{Updates a Text Property}
\usage{
update_text(secret, id, property_name, value)
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
@param secret API token
@param id Page id to be updated
@param property_name name of property to update (should be a text type property)
@param value value to update
}
