% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePageCover.R
\name{updatePageCover}
\alias{updatePageCover}
\title{Updates a Page Cover}
\usage{
updatePageCover(secret, id, cover_url)
}
\arguments{
\item{secret}{Notion API token}

\item{id}{page id to be updated}

\item{cover_url}{url of cover to be update}
}
\value{
list of response

list
}
\description{
Updates a page cover to the url specified
}
\author{
Eduardo Flores
}
