% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSelect.R
\name{updateSelect}
\alias{updateSelect}
\title{Updates a Select Property}
\usage{
updateSelect(secret, id, property_name, value)
}
\arguments{
\item{secret}{API token}

\item{id}{Page id to be updated}

\item{property_name}{name of property to update (should be a select type property)}

\item{value}{value to update}
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
}
