% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateText.R
\name{updateText}
\alias{updateText}
\title{Updates a Text Property}
\usage{
updateText(secret, id, property_name, value)
}
\arguments{
\item{secret}{API token}

\item{id}{Page id to be updated}

\item{property_name}{name of property to update (should be a text type property)}

\item{value}{value to update}
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
}
