% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTitle.R
\name{updateTitle}
\alias{updateTitle}
\title{Updates a Title Property}
\usage{
updateTitle(secret, id, property_name, value)
}
\arguments{
\item{secret}{API token}

\item{id}{Page id to be updated}

\item{property_name}{name of property to update (should be the title property of database)}

\item{value}{value to update}
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
}
