% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{users}
\title{Get users}
\usage{
users(key)
}
\arguments{
\item{key}{Notion access key}
}
\value{
A list of user objects
}
\description{
This endpoint retrieves all users with authorized access to a workspace
}
