% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBlockH1.R
\name{addBlockH1}
\alias{addBlockH1}
\title{Adds an H1 (heading one) Block to a Page}
\usage{
addBlockH1(secret, id, content, toggle = FALSE)
}
\description{
Id refers to a page id, content should be only text.
@param secret API token
@param id Page id where block will be appended
@param content content to append as H1
@param toggle defaults to FALSE. If TRUE, will create an H1 Toggle.
}
