% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBlockParagraph.R
\name{addBlockParagraph}
\alias{addBlockParagraph}
\title{Adds a Paragraph Block to a Page}
\usage{
addBlockParagraph(secret, id, content)
}
\arguments{
\item{secret}{API token}

\item{id}{Page id where block will be appended}

\item{content}{content to append}
}
\description{
Id refers to a page id, content should be only text. HTML will export to text.
}
