% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEmoji.R
\name{updateEmoji}
\alias{updateEmoji}
\title{Updates an Emoji Property of a page}
\usage{
updateEmoji(secret, id, emoji)
}
\arguments{
\item{secret}{API token}

\item{id}{Page id to be updated}

\item{emoji}{emoji to update to}
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
}
