% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMultiSelect.R
\name{updateMultiSelect}
\alias{updateMultiSelect}
\title{Updates a Select Property}
\usage{
updateMultiSelect(secret, id, property_name, value)
}
\arguments{
\item{secret}{API token}

\item{id}{Page id to be updated}

\item{property_name}{name of property to update (should be a multiselect type property)}

\item{value}{value(s) to update. Could be 1 value or multiple, created with c().}
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
}
\details{
This will rewrite whatever is already in the property. It will NOT append another select.
}
