% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRelationship.R
\name{updateRelationship}
\alias{updateRelationship}
\title{Updates (adds) a relationship to a page id}
\usage{
updateRelationship(secret, id, property_name, value)
}
\arguments{
\item{secret}{API token}

\item{id}{Page id to be updated}

\item{property_name}{name of property to update (should be a relationship type property)}

\item{value}{value to update (should be a unique page id)}
}
\description{
Id refers to a page in a database, and should be normalized using normalizeChromeId().
}
